requirements.txt
python-dotenv~=0.19.2
aiohttp~=3.7.4
discord.py~=2.0.1
typing_extensions~=4.7.1